requirements.txt
Django==5.1.5  
django-anymail==10.2  
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-cors-headers==3.14.0
django-environ==0.9.0
django-import-export==3.2.0
django-jazzmin==2.6.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
humanize==4.6.0
requests==2.31.0
shortuuid==1.0.11
django-cors-headers
whitenoise==6.8.2
gunicorn==23.0.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
setuptools